Initial HTML Skeletons
Here's the skeleton for the HTML Files. I added a bp_results and cvp_results page to view the individual analysis for each product. Get & Update functions need to be implemented for database integration and the hard-coded example data needs to be removed. SUMMARY: TEMPORARY WITH UPDATES TO BE ADDED LATER THIS WEEK! :)