Added parts for Flask
I'm still not 100% comfortable with Flask and whatnot, so any and all edits are welcome. Scrap it all if you have to. Some of the variables I was working with, I don't think, have any definitions in the files so far (for example: Delete (data), Edit (users)), so let me know what you guys see and think, I could be completely wrong too.